remove require from base-server
